browser: split up strokes code
