+++ packages/browser/src/keyboard/index.ts
@@ -1,0 +1,9 @@
+export { characterToCode } from "./keys";
+
+export {
+  serializeStrokes,
+  stringToStrokes,
+  valueToStrokes
+} from "./serializeStrokes";
+
+export { Stroke, StrokeType } from "./Stroke";
