+++ packages/browser/tests/keyboard/keys.test.ts
@@ -1,0 +1,11 @@
+import { characterToCode } from "../../src/keyboard/keys";
+
+describe("characterToCode", () => {
+  it("converts a character to it's USKeyboard code", () => {
+    expect(characterToCode("S")).toEqual("KeyS");
+  });
+
+  it("converts a non-USKeyboard character to null", () => {
+    expect(characterToCode("嗨")).toEqual(null);
+  });
+});
